+++ src/engine/impl/gltf_parser.h
@@ -16,7 +16,7 @@ struct GeometryInfo
 {
     engine_vertex_attributes_layout_t vertex_laytout{};
     std::vector<std::byte> vertex_data;
-    std::size_t vertex_count = 0;
+    std::int32_t vertex_count = 0;
     std::vector<std::uint32_t> indicies;
     std::int32_t material_index = -1;
 };
